sparse: represent matrix row count as 'size_t'
This is part of a series of changes to pervasively use `size_t` for
counting. This change pragmatically casts back to `int` for interacting
with legacy code. This may be the only feasible way to achieve this kind
of progress. This change squashes 26 -Wsign-conversion warnings.

+++ cmd/tools/matrix_market.c
@@ -86,8 +86,8 @@ SparseMatrix SparseMatrix_import_matrix_market(FILE *f) {
   }
   vp = val;
 
-  A = SparseMatrix_from_coordinate_arrays(nz, m, n, I, J, vp, MATRIX_TYPE_REAL,
-                                          sizeof(double));
+  A = SparseMatrix_from_coordinate_arrays(nz, (size_t)m, n, I, J, vp,
+                                          MATRIX_TYPE_REAL, sizeof(double));
 done:
   free(I);
   free(J);
